requirements.txt
Flask==2.3.3
flask-cors==3.0.10
numpy==1.25.2
pandas==2.1.1
scikit-learn==1.3.0
joblib==1.3.2
tensorflow-cpu==2.14.0
gunicorn==20.1.0
